requirements.txt
pygame
keyboard
ffmpeg-python
scipy==1.10.1
mutagen==1.46.0
pynput==1.7.6
pydub==0.25.1
click==8.1.3
textual==0.23.0
pyyaml==6.0
dotmap==1.3.30
